tests/wpt/tests/mathml/relations/html5-tree/a-hreflang-setter.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>MathMLAnchorElement.hreflang setter</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <math>
      <a id="test" href="a">Example Link</a>
    </math>
  </body>
  <script>
      test(function() {
          var a = document.getElementById("test");
          a.hreflang = "fr"
          assert_equals(a.hreflang, "fr");
          assert_equals(a.getAttribute('hreflang'), "fr");
      }, "Test anchor's hreflang setter");
  </script>
</html>
